checkers/requirements.txt
requests==2.32.3
python-socketio==5.11.4
websocket-client==1.8.0
checklib==0.7.0
pillow == 11
pytesseract
